requirements.txt
python-dotenv==1.0.1
beautifulsoup4==4.12.3
aiohttp>=3.9.0
playwright>=1.40.0
